osa 4:n tehtävät tehty.

+++ blogilista/controllers/blogs.js
@@ -1,44 +1,49 @@
-const blogsRouter = require('express').Router()
-const Blog = require('../models/blog')
+const blogsRouter = require("express").Router();
+const Blog = require("../models/blog");
+const middleWare = require("../utils/middleware")
 
-blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
-  response.json(blogs.map(blog => blog.toJSON()))
-})
-  
-blogsRouter.post('/', async (request, response, next) => {
-  const blog = new Blog(request.body)
+blogsRouter.get("/", async (request, response) => {
+  const blogs = await Blog.find({}).populate("user", { username: 1, name: 1 });
 
-  try { 
-    const savedBlog = await blog.save()
-    response.json(savedBlog.toJSON())
-  } catch(exception) {
-    next(exception)
+  response.json(blogs.map((blog) => blog.toJSON()));
+});
+
+blogsRouter.post("/", middleWare.tokenExtractor, middleWare.userExtractor, async (request, response) => {
+  const body = request.body;
+
+  const blog = new Blog({
+    title: body.title,
+    author: body.author,
+    url: body.url,
+    user: request.user._id,
+  });
+
+  const savedBlog = await blog.save();
+  request.user.blogs = request.user.blogs.concat(savedBlog._id);
+  await request.user.save();
+  return response.json(savedBlog.toJSON());
+});
+
+blogsRouter.delete("/:id", middleWare.tokenExtractor, middleWare.userExtractor, async (request, response) => {
+
+  const blog = await Blog.findById(request.params.id)
+  if (blog.user.toString() === request.user._id.toString()) {
+    await Blog.findByIdAndRemove(request.params.id);
+    response.status(204).end();
   }
-})
+  return response.status(400).json({ error: "No permission to do that" });
+});
 
-blogsRouter.delete('/:id', async (request, response, next) => {
-  try { 
-    await Blog.findByIdAndRemove(request.params.id)
-    response.status(204).end()
-  } catch(exception) {
-    next(exception)
-  }
-})
-
-blogsRouter.put('/:id', async (request, response, next) => {
-  const body = await request.body
+blogsRouter.put("/:id", async (request, response) => {
+  const body = await request.body;
 
   const updatedBlog = {
-    likes: body.likes
-  }
+    likes: body.likes,
+  };
+  const b = await Blog.findByIdAndUpdate(request.params.id, updatedBlog, {
+    new: true,
+  });
+  response.json(b);
+});
 
-  try {
-    const b = await Blog.findByIdAndUpdate(request.params.id, updatedBlog, { new: true })
-    response.json(b)
-  } catch(exception) {
-    next(exception)
-  }
-})
-
-module.exports = blogsRouter
+module.exports = blogsRouter;
